+++ src/context/TokenContext.ts
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const TokenContext = createContext<string | null>(null);
+
+export default TokenContext;
