--- src/hook/useBlockLoginUser.ts
@@ -1,14 +1,0 @@
-import { useEffect } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { getAuthToken } from '../utils/token/token.util';
-
-export default function useBlockLoginUser() {
-  const navigate = useNavigate();
-
-  useEffect(() => {
-    const token = getAuthToken();
-    if (token) {
-      navigate('/');
-    }
-  }, [navigate]);
-}
